+++ src/video_core/CMakeLists.txt
@@ -23,6 +23,8 @@ add_library(video_core STATIC
     regs_texturing.h
     renderer_base.cpp
     renderer_base.h
+    rasterizer_cache/cached_surface.cpp
+    rasterizer_cache/cached_surface.h
     rasterizer_cache/morton_swizzle.h
     rasterizer_cache/pixel_format.h
     rasterizer_cache/rasterizer_cache.cpp
@@ -32,6 +34,8 @@ add_library(video_core STATIC
     rasterizer_cache/rasterizer_cache_utils.h
     rasterizer_cache/surface_params.cpp
     rasterizer_cache/surface_params.h
+    rasterizer_cache/texture_runtime.cpp
+    rasterizer_cache/texture_runtime.h
     renderer_opengl/frame_dumper_opengl.cpp
     renderer_opengl/frame_dumper_opengl.h
     renderer_opengl/gl_rasterizer.cpp
