+++ src/video_core/rasterizer_cache/rasterizer_cache.cpp
@@ -4,21 +4,73 @@
 
 #include <optional>
 #include <boost/range/iterator_range.hpp>
+#include "common/alignment.h"
 #include "common/logging/log.h"
 #include "common/microprofile.h"
 #include "common/scope_exit.h"
-#include "common/texture.h"
-#include "core/core.h"
-#include "core/hle/kernel/process.h"
 #include "video_core/pica_state.h"
-#include "video_core/rasterizer_cache/morton_swizzle.h"
 #include "video_core/rasterizer_cache/rasterizer_cache.h"
+#include "video_core/renderer_opengl/texture_filters/texture_filterer.h"
+#include "video_core/renderer_opengl/texture_downloader_es.h"
 #include "video_core/renderer_opengl/gl_format_reinterpreter.h"
 #include "video_core/renderer_opengl/gl_state.h"
-#include "video_core/renderer_opengl/texture_downloader_es.h"
-#include "video_core/renderer_opengl/texture_filters/texture_filterer.h"
+#include "video_core/renderer_opengl/gl_vars.h"
 
 namespace OpenGL {
+
+// TODO: Deduplicate this
+static Aspect ToAspect(SurfaceType type) {
+    switch (type) {
+    case SurfaceType::Color:
+    case SurfaceType::Texture:
+    case SurfaceType::Fill:
+        return Aspect::Color;
+    case SurfaceType::Depth:
+        return Aspect::Depth;
+    case SurfaceType::DepthStencil:
+        return Aspect::DepthStencil;
+    default:
+        LOG_CRITICAL(Render_OpenGL, "Unknown SurfaceType {}", type);
+        UNREACHABLE();
+    }
+
+    return Aspect::Color;
+}
+
+static ClearValue ToClearValue(Aspect aspect, PixelFormat format, const u8* fill_data) {
+    ClearValue result{};
+    switch (aspect) {
+    case Aspect::Color: {
+        Pica::Texture::TextureInfo tex_info{};
+        tex_info.format = static_cast<Pica::TexturingRegs::TextureFormat>(format);
+
+        Common::Vec4<u8> color = Pica::Texture::LookupTexture(fill_data, 0, 0, tex_info);
+        result.color = color / 255.f;
+        break;
+    }
+    case Aspect::Depth: {
+        u32 depth_uint = 0;
+        if (format == PixelFormat::D16) {
+            std::memcpy(&depth_uint, fill_data, 2);
+            result.depth = depth_uint / 65535.0f; // 2^16 - 1
+        } else if (format == PixelFormat::D24) {
+            std::memcpy(&depth_uint, fill_data, 3);
+            result.depth = depth_uint / 16777215.0f; // 2^24 - 1
+        }
+        break;
+    }
+    case Aspect::DepthStencil: {
+        u32 clear_value_uint;
+        std::memcpy(&clear_value_uint, fill_data, sizeof(u32));
+
+        result.depth = (clear_value_uint & 0xFFFFFF) / 16777215.0f; // 2^24 - 1
+        result.stencil = (clear_value_uint >> 24);
+        break;
+    }
+    }
+
+    return result;
+}
 
 template <typename Map, typename Interval>
 static constexpr auto RangeFromInterval(Map& map, const Interval& interval) {
@@ -96,192 +148,6 @@ static void AllocateTextureCube(GLuint texture, const FormatTuple& format_tuple,
     cur_state.Apply();
 }
 
-static bool BlitTextures(GLuint src_tex, const Common::Rectangle<u32>& src_rect, GLuint dst_tex,
-                         const Common::Rectangle<u32>& dst_rect, SurfaceType type,
-                         GLuint read_fb_handle, GLuint draw_fb_handle) {
-    OpenGLState prev_state = OpenGLState::GetCurState();
-    SCOPE_EXIT({ prev_state.Apply(); });
-
-    OpenGLState state;
-    state.draw.read_framebuffer = read_fb_handle;
-    state.draw.draw_framebuffer = draw_fb_handle;
-    state.Apply();
-
-    u32 buffers = 0;
-
-    if (type == SurfaceType::Color || type == SurfaceType::Texture) {
-        glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, src_tex,
-                               0);
-        glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT, GL_TEXTURE_2D, 0,
-                               0);
-
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, dst_tex,
-                               0);
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT, GL_TEXTURE_2D, 0,
-                               0);
-
-        buffers = GL_COLOR_BUFFER_BIT;
-    } else if (type == SurfaceType::Depth) {
-        glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, 0, 0);
-        glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_TEXTURE_2D, src_tex, 0);
-        glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_STENCIL_ATTACHMENT, GL_TEXTURE_2D, 0, 0);
-
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, 0, 0);
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_TEXTURE_2D, dst_tex, 0);
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_STENCIL_ATTACHMENT, GL_TEXTURE_2D, 0, 0);
-
-        buffers = GL_DEPTH_BUFFER_BIT;
-    } else if (type == SurfaceType::DepthStencil) {
-        glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, 0, 0);
-        glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT, GL_TEXTURE_2D,
-                               src_tex, 0);
-
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, 0, 0);
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT, GL_TEXTURE_2D,
-                               dst_tex, 0);
-
-        buffers = GL_DEPTH_BUFFER_BIT | GL_STENCIL_BUFFER_BIT;
-    }
-
-    // TODO (wwylele): use GL_NEAREST for shadow map texture
-    // Note: shadow map is treated as RGBA8 format in PICA, as well as in the rasterizer cache, but
-    // doing linear intepolation componentwise would cause incorrect value. However, for a
-    // well-programmed game this code path should be rarely executed for shadow map with
-    // inconsistent scale.
-    glBlitFramebuffer(src_rect.left, src_rect.bottom, src_rect.right, src_rect.top, dst_rect.left,
-                      dst_rect.bottom, dst_rect.right, dst_rect.top, buffers,
-                      buffers == GL_COLOR_BUFFER_BIT ? GL_LINEAR : GL_NEAREST);
-
-    return true;
-}
-
-static bool FillSurface(const Surface& surface, const u8* fill_data,
-                        const Common::Rectangle<u32>& fill_rect, GLuint draw_fb_handle) {
-    OpenGLState prev_state = OpenGLState::GetCurState();
-    SCOPE_EXIT({ prev_state.Apply(); });
-
-    OpenGLState state;
-    state.scissor.enabled = true;
-    state.scissor.x = static_cast<GLint>(fill_rect.left);
-    state.scissor.y = static_cast<GLint>(fill_rect.bottom);
-    state.scissor.width = static_cast<GLsizei>(fill_rect.GetWidth());
-    state.scissor.height = static_cast<GLsizei>(fill_rect.GetHeight());
-
-    state.draw.draw_framebuffer = draw_fb_handle;
-    state.Apply();
-
-    surface->InvalidateAllWatcher();
-
-    if (surface->type == SurfaceType::Color || surface->type == SurfaceType::Texture) {
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D,
-                               surface->texture.handle, 0);
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT, GL_TEXTURE_2D, 0,
-                               0);
-
-        Pica::Texture::TextureInfo tex_info{};
-        tex_info.format = static_cast<Pica::TexturingRegs::TextureFormat>(surface->pixel_format);
-        Common::Vec4<u8> color = Pica::Texture::LookupTexture(fill_data, 0, 0, tex_info);
-
-        std::array<GLfloat, 4> color_values = {color.x / 255.f, color.y / 255.f, color.z / 255.f,
-                                               color.w / 255.f};
-
-        state.color_mask.red_enabled = GL_TRUE;
-        state.color_mask.green_enabled = GL_TRUE;
-        state.color_mask.blue_enabled = GL_TRUE;
-        state.color_mask.alpha_enabled = GL_TRUE;
-        state.Apply();
-        glClearBufferfv(GL_COLOR, 0, &color_values[0]);
-    } else if (surface->type == SurfaceType::Depth) {
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, 0, 0);
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_TEXTURE_2D,
-                               surface->texture.handle, 0);
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_STENCIL_ATTACHMENT, GL_TEXTURE_2D, 0, 0);
-
-        u32 value_32bit = 0;
-        GLfloat value_float;
-
-        if (surface->pixel_format == PixelFormat::D16) {
-            std::memcpy(&value_32bit, fill_data, 2);
-            value_float = value_32bit / 65535.0f; // 2^16 - 1
-        } else if (surface->pixel_format == PixelFormat::D24) {
-            std::memcpy(&value_32bit, fill_data, 3);
-            value_float = value_32bit / 16777215.0f; // 2^24 - 1
-        }
-
-        state.depth.write_mask = GL_TRUE;
-        state.Apply();
-        glClearBufferfv(GL_DEPTH, 0, &value_float);
-    } else if (surface->type == SurfaceType::DepthStencil) {
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, 0, 0);
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT, GL_TEXTURE_2D,
-                               surface->texture.handle, 0);
-
-        u32 value_32bit;
-        std::memcpy(&value_32bit, fill_data, sizeof(u32));
-
-        GLfloat value_float = (value_32bit & 0xFFFFFF) / 16777215.0f; // 2^24 - 1
-        GLint value_int = (value_32bit >> 24);
-
-        state.depth.write_mask = GL_TRUE;
-        state.stencil.write_mask = -1;
-        state.Apply();
-        glClearBufferfi(GL_DEPTH_STENCIL, 0, value_float, value_int);
-    }
-    return true;
-}
-
-CachedSurface::~CachedSurface() {
-    if (texture.handle) {
-        auto tag = is_custom ? HostTextureTag{GetFormatTuple(PixelFormat::RGBA8),
-                                              custom_tex_info.width, custom_tex_info.height}
-                             : HostTextureTag{GetFormatTuple(pixel_format), GetScaledWidth(),
-                                              GetScaledHeight()};
-
-        owner.host_texture_recycler.emplace(tag, std::move(texture));
-    }
-}
-
-bool CachedSurface::CanFill(const SurfaceParams& dest_surface,
-                            SurfaceInterval fill_interval) const {
-    if (type == SurfaceType::Fill && IsRegionValid(fill_interval) &&
-        boost::icl::first(fill_interval) >= addr &&
-        boost::icl::last_next(fill_interval) <= end && // dest_surface is within our fill range
-        dest_surface.FromInterval(fill_interval).GetInterval() ==
-            fill_interval) { // make sure interval is a rectangle in dest surface
-        if (fill_size * 8 != dest_surface.GetFormatBpp()) {
-            // Check if bits repeat for our fill_size
-            const u32 dest_bytes_per_pixel = std::max(dest_surface.GetFormatBpp() / 8, 1u);
-            std::vector<u8> fill_test(fill_size * dest_bytes_per_pixel);
-
-            for (u32 i = 0; i < dest_bytes_per_pixel; ++i)
-                std::memcpy(&fill_test[i * fill_size], &fill_data[0], fill_size);
-
-            for (u32 i = 0; i < fill_size; ++i)
-                if (std::memcmp(&fill_test[dest_bytes_per_pixel * i], &fill_test[0],
-                                dest_bytes_per_pixel) != 0)
-                    return false;
-
-            if (dest_surface.GetFormatBpp() == 4 && (fill_test[0] & 0xF) != (fill_test[0] >> 4))
-                return false;
-        }
-        return true;
-    }
-    return false;
-}
-
-bool CachedSurface::CanCopy(const SurfaceParams& dest_surface,
-                            SurfaceInterval copy_interval) const {
-    SurfaceParams subrect_params = dest_surface.FromInterval(copy_interval);
-    ASSERT(subrect_params.GetInterval() == copy_interval);
-    if (CanSubRect(subrect_params))
-        return true;
-
-    if (CanFill(dest_surface, copy_interval))
-        return true;
-
-    return false;
-}
-
 MICROPROFILE_DEFINE(OpenGL_CopySurface, "OpenGL", "CopySurface", MP_RGB(128, 192, 64));
 void RasterizerCacheOpenGL::CopySurface(const Surface& src_surface, const Surface& dst_surface,
                                         SurfaceInterval copy_interval) {
@@ -289,10 +155,10 @@ void RasterizerCacheOpenGL::CopySurface(const Surface& src_surface, const Surfac
 
     SurfaceParams subrect_params = dst_surface->FromInterval(copy_interval);
     ASSERT(subrect_params.GetInterval() == copy_interval);
-
     ASSERT(src_surface != dst_surface);
 
     // This is only called when CanCopy is true, no need to run checks here
+    const Aspect aspect = ToAspect(dst_surface->type);
     if (src_surface->type == SurfaceType::Fill) {
         // FillSurface needs a 4 bytes buffer
         const u32 fill_offset =
@@ -300,441 +166,28 @@ void RasterizerCacheOpenGL::CopySurface(const Surface& src_surface, const Surfac
         std::array<u8, 4> fill_buffer;
 
         u32 fill_buff_pos = fill_offset;
-        for (int i : {0, 1, 2, 3})
+        for (std::size_t i = 0; i < fill_buffer.size(); i++) {
             fill_buffer[i] = src_surface->fill_data[fill_buff_pos++ % src_surface->fill_size];
+        }
 
-        FillSurface(dst_surface, &fill_buffer[0], dst_surface->GetScaledSubRect(subrect_params),
-                    draw_framebuffer.handle);
+        const auto clear_rect = dst_surface->GetScaledSubRect(subrect_params);
+        const ClearValue clear_value = ToClearValue(aspect, dst_surface->pixel_format,
+                                                    fill_buffer.data());
+
+        runtime.ClearTexture(dst_surface->texture, {aspect, clear_rect}, clear_value);
         return;
     }
+
     if (src_surface->CanSubRect(subrect_params)) {
-        BlitTextures(src_surface->texture.handle, src_surface->GetScaledSubRect(subrect_params),
-                     dst_surface->texture.handle, dst_surface->GetScaledSubRect(subrect_params),
-                     src_surface->type, read_framebuffer.handle, draw_framebuffer.handle);
+        const auto src_rect = src_surface->GetScaledSubRect(subrect_params);
+        const auto dst_rect = dst_surface->GetScaledSubRect(subrect_params);
+
+        runtime.BlitTextures(src_surface->texture, {aspect, src_rect},
+                             dst_surface->texture, {aspect, dst_rect});
         return;
     }
+
     UNREACHABLE();
-}
-
-MICROPROFILE_DEFINE(OpenGL_SurfaceLoad, "OpenGL", "Surface Load", MP_RGB(128, 192, 64));
-void CachedSurface::LoadGLBuffer(PAddr load_start, PAddr load_end) {
-    ASSERT(type != SurfaceType::Fill);
-    const bool need_swap =
-        GLES && (pixel_format == PixelFormat::RGBA8 || pixel_format == PixelFormat::RGB8);
-
-    const u8* const texture_src_data = VideoCore::g_memory->GetPhysicalPointer(addr);
-    if (texture_src_data == nullptr)
-        return;
-
-    if (gl_buffer.empty()) {
-        gl_buffer.resize(width * height * GetBytesPerPixel(pixel_format));
-    }
-
-    // TODO: Should probably be done in ::Memory:: and check for other regions too
-    if (load_start < Memory::VRAM_VADDR_END && load_end > Memory::VRAM_VADDR_END)
-        load_end = Memory::VRAM_VADDR_END;
-
-    if (load_start < Memory::VRAM_VADDR && load_end > Memory::VRAM_VADDR)
-        load_start = Memory::VRAM_VADDR;
-
-    MICROPROFILE_SCOPE(OpenGL_SurfaceLoad);
-
-    ASSERT(load_start >= addr && load_end <= end);
-    const u32 start_offset = load_start - addr;
-
-    if (!is_tiled) {
-        ASSERT(type == SurfaceType::Color);
-        if (need_swap) {
-            // TODO(liushuyu): check if the byteswap here is 100% correct
-            // cannot fully test this
-            if (pixel_format == PixelFormat::RGBA8) {
-                for (std::size_t i = start_offset; i < load_end - addr; i += 4) {
-                    gl_buffer[i] = texture_src_data[i + 3];
-                    gl_buffer[i + 1] = texture_src_data[i + 2];
-                    gl_buffer[i + 2] = texture_src_data[i + 1];
-                    gl_buffer[i + 3] = texture_src_data[i];
-                }
-            } else if (pixel_format == PixelFormat::RGB8) {
-                for (std::size_t i = start_offset; i < load_end - addr; i += 3) {
-                    gl_buffer[i] = texture_src_data[i + 2];
-                    gl_buffer[i + 1] = texture_src_data[i + 1];
-                    gl_buffer[i + 2] = texture_src_data[i];
-                }
-            }
-        } else {
-            std::memcpy(&gl_buffer[start_offset], texture_src_data + start_offset,
-                        load_end - load_start);
-        }
-    } else {
-        if (type == SurfaceType::Texture) {
-            Pica::Texture::TextureInfo tex_info{};
-            tex_info.width = width;
-            tex_info.height = height;
-            tex_info.format = static_cast<Pica::TexturingRegs::TextureFormat>(pixel_format);
-            tex_info.SetDefaultStride();
-            tex_info.physical_address = addr;
-
-            const SurfaceInterval load_interval(load_start, load_end);
-            const auto rect = GetSubRect(FromInterval(load_interval));
-            ASSERT(FromInterval(load_interval).GetInterval() == load_interval);
-
-            for (unsigned y = rect.bottom; y < rect.top; ++y) {
-                for (unsigned x = rect.left; x < rect.right; ++x) {
-                    auto vec4 =
-                        Pica::Texture::LookupTexture(texture_src_data, x, height - 1 - y, tex_info);
-                    const std::size_t offset = (x + (width * y)) * 4;
-                    std::memcpy(&gl_buffer[offset], vec4.AsArray(), 4);
-                }
-            }
-        } else {
-            morton_to_gl_fns[static_cast<std::size_t>(pixel_format)](stride, height, &gl_buffer[0],
-                                                                     addr, load_start, load_end);
-        }
-    }
-}
-
-MICROPROFILE_DEFINE(OpenGL_SurfaceFlush, "OpenGL", "Surface Flush", MP_RGB(128, 192, 64));
-void CachedSurface::FlushGLBuffer(PAddr flush_start, PAddr flush_end) {
-    u8* const dst_buffer = VideoCore::g_memory->GetPhysicalPointer(addr);
-    if (dst_buffer == nullptr)
-        return;
-
-    ASSERT(gl_buffer.size() == width * height * GetBytesPerPixel(pixel_format));
-
-    // TODO: Should probably be done in ::Memory:: and check for other regions too
-    // same as loadglbuffer()
-    if (flush_start < Memory::VRAM_VADDR_END && flush_end > Memory::VRAM_VADDR_END)
-        flush_end = Memory::VRAM_VADDR_END;
-
-    if (flush_start < Memory::VRAM_VADDR && flush_end > Memory::VRAM_VADDR)
-        flush_start = Memory::VRAM_VADDR;
-
-    MICROPROFILE_SCOPE(OpenGL_SurfaceFlush);
-
-    ASSERT(flush_start >= addr && flush_end <= end);
-    const u32 start_offset = flush_start - addr;
-    const u32 end_offset = flush_end - addr;
-
-    if (type == SurfaceType::Fill) {
-        const u32 coarse_start_offset = start_offset - (start_offset % fill_size);
-        const u32 backup_bytes = start_offset % fill_size;
-        std::array<u8, 4> backup_data;
-        if (backup_bytes)
-            std::memcpy(&backup_data[0], &dst_buffer[coarse_start_offset], backup_bytes);
-
-        for (u32 offset = coarse_start_offset; offset < end_offset; offset += fill_size) {
-            std::memcpy(&dst_buffer[offset], &fill_data[0],
-                        std::min(fill_size, end_offset - offset));
-        }
-
-        if (backup_bytes)
-            std::memcpy(&dst_buffer[coarse_start_offset], &backup_data[0], backup_bytes);
-    } else if (!is_tiled) {
-        ASSERT(type == SurfaceType::Color);
-        if (pixel_format == PixelFormat::RGBA8 && GLES) {
-            for (std::size_t i = start_offset; i < flush_end - addr; i += 4) {
-                dst_buffer[i] = gl_buffer[i + 3];
-                dst_buffer[i + 1] = gl_buffer[i + 2];
-                dst_buffer[i + 2] = gl_buffer[i + 1];
-                dst_buffer[i + 3] = gl_buffer[i];
-            }
-        } else if (pixel_format == PixelFormat::RGB8 && GLES) {
-            for (std::size_t i = start_offset; i < flush_end - addr; i += 3) {
-                dst_buffer[i] = gl_buffer[i + 2];
-                dst_buffer[i + 1] = gl_buffer[i + 1];
-                dst_buffer[i + 2] = gl_buffer[i];
-            }
-        } else {
-            std::memcpy(dst_buffer + start_offset, &gl_buffer[start_offset],
-                        flush_end - flush_start);
-        }
-    } else {
-        gl_to_morton_fns[static_cast<std::size_t>(pixel_format)](stride, height, &gl_buffer[0],
-                                                                 addr, flush_start, flush_end);
-    }
-}
-
-bool CachedSurface::LoadCustomTexture(u64 tex_hash) {
-    auto& custom_tex_cache = Core::System::GetInstance().CustomTexCache();
-    const auto& image_interface = Core::System::GetInstance().GetImageInterface();
-
-    if (custom_tex_cache.IsTextureCached(tex_hash)) {
-        custom_tex_info = custom_tex_cache.LookupTexture(tex_hash);
-        return true;
-    }
-
-    if (!custom_tex_cache.CustomTextureExists(tex_hash)) {
-        return false;
-    }
-
-    const auto& path_info = custom_tex_cache.LookupTexturePathInfo(tex_hash);
-    if (!image_interface->DecodePNG(custom_tex_info.tex, custom_tex_info.width,
-                                    custom_tex_info.height, path_info.path)) {
-        LOG_ERROR(Render_OpenGL, "Failed to load custom texture {}", path_info.path);
-        return false;
-    }
-
-    const std::bitset<32> width_bits(custom_tex_info.width);
-    const std::bitset<32> height_bits(custom_tex_info.height);
-    if (width_bits.count() != 1 || height_bits.count() != 1) {
-        LOG_ERROR(Render_OpenGL, "Texture {} size is not a power of 2", path_info.path);
-        return false;
-    }
-
-    LOG_DEBUG(Render_OpenGL, "Loaded custom texture from {}", path_info.path);
-    Common::FlipRGBA8Texture(custom_tex_info.tex, custom_tex_info.width, custom_tex_info.height);
-    custom_tex_cache.CacheTexture(tex_hash, custom_tex_info.tex, custom_tex_info.width,
-                                  custom_tex_info.height);
-    return true;
-}
-
-void CachedSurface::DumpTexture(GLuint target_tex, u64 tex_hash) {
-    // Make sure the texture size is a power of 2
-    // If not, the surface is actually a framebuffer
-    std::bitset<32> width_bits(width);
-    std::bitset<32> height_bits(height);
-    if (width_bits.count() != 1 || height_bits.count() != 1) {
-        LOG_WARNING(Render_OpenGL, "Not dumping {:016X} because size isn't a power of 2 ({}x{})",
-                    tex_hash, width, height);
-        return;
-    }
-
-    // Dump texture to RGBA8 and encode as PNG
-    const auto& image_interface = Core::System::GetInstance().GetImageInterface();
-    auto& custom_tex_cache = Core::System::GetInstance().CustomTexCache();
-    std::string dump_path =
-        fmt::format("{}textures/{:016X}/", FileUtil::GetUserPath(FileUtil::UserPath::DumpDir),
-                    Core::System::GetInstance().Kernel().GetCurrentProcess()->codeset->program_id);
-    if (!FileUtil::CreateFullPath(dump_path)) {
-        LOG_ERROR(Render, "Unable to create {}", dump_path);
-        return;
-    }
-
-    dump_path += fmt::format("tex1_{}x{}_{:016X}_{}.png", width, height, tex_hash, pixel_format);
-    if (!custom_tex_cache.IsTextureDumped(tex_hash) && !FileUtil::Exists(dump_path)) {
-        custom_tex_cache.SetTextureDumped(tex_hash);
-
-        LOG_INFO(Render_OpenGL, "Dumping texture to {}", dump_path);
-        std::vector<u8> decoded_texture;
-        decoded_texture.resize(width * height * 4);
-        OpenGLState state = OpenGLState::GetCurState();
-        GLuint old_texture = state.texture_units[0].texture_2d;
-        state.Apply();
-        /*
-           GetTexImageOES is used even if not using OpenGL ES to work around a small issue that
-           happens if using custom textures with texture dumping at the same.
-           Let's say there's 2 textures that are both 32x32 and one of them gets replaced with a
-           higher quality 256x256 texture. If the 256x256 texture is displayed first and the
-           32x32 texture gets uploaded to the same underlying OpenGL texture, the 32x32 texture
-           will appear in the corner of the 256x256 texture. If texture dumping is enabled and
-           the 32x32 is undumped, Citra will attempt to dump it. Since the underlying OpenGL
-           texture is still 256x256, Citra crashes because it thinks the texture is only 32x32.
-           GetTexImageOES conveniently only dumps the specified region, and works on both
-           desktop and ES.
-        */
-        // if the backend isn't OpenGL ES, this won't be initialized yet
-        if (!owner.texture_downloader_es)
-            owner.texture_downloader_es = std::make_unique<TextureDownloaderES>(false);
-        owner.texture_downloader_es->GetTexImage(GL_TEXTURE_2D, 0, GL_RGBA, GL_UNSIGNED_BYTE,
-                                                 height, width, &decoded_texture[0]);
-        state.texture_units[0].texture_2d = old_texture;
-        state.Apply();
-        Common::FlipRGBA8Texture(decoded_texture, width, height);
-        if (!image_interface->EncodePNG(dump_path, decoded_texture, width, height))
-            LOG_ERROR(Render_OpenGL, "Failed to save decoded texture");
-    }
-}
-
-MICROPROFILE_DEFINE(OpenGL_TextureUL, "OpenGL", "Texture Upload", MP_RGB(128, 192, 64));
-void CachedSurface::UploadGLTexture(Common::Rectangle<u32> rect, GLuint read_fb_handle,
-                                    GLuint draw_fb_handle) {
-    if (type == SurfaceType::Fill)
-        return;
-
-    MICROPROFILE_SCOPE(OpenGL_TextureUL);
-
-    ASSERT(gl_buffer.size() == width * height * GetBytesPerPixel(pixel_format));
-
-    u64 tex_hash = 0;
-
-    if (Settings::values.dump_textures || Settings::values.custom_textures) {
-        tex_hash = Common::ComputeHash64(gl_buffer.data(), gl_buffer.size());
-    }
-
-    if (Settings::values.custom_textures) {
-        is_custom = LoadCustomTexture(tex_hash);
-    }
-
-    // Load data from memory to the surface
-    GLint x0 = static_cast<GLint>(rect.left);
-    GLint y0 = static_cast<GLint>(rect.bottom);
-    std::size_t buffer_offset = (y0 * stride + x0) * GetBytesPerPixel(pixel_format);
-
-    const FormatTuple& tuple = GetFormatTuple(pixel_format);
-    GLuint target_tex = texture.handle;
-
-    // If not 1x scale, create 1x texture that we will blit from to replace texture subrect in
-    // surface
-    OGLTexture unscaled_tex;
-    if (res_scale != 1) {
-        x0 = 0;
-        y0 = 0;
-
-        if (is_custom) {
-            unscaled_tex = owner.AllocateSurfaceTexture(
-                GetFormatTuple(PixelFormat::RGBA8), custom_tex_info.width, custom_tex_info.height);
-        } else {
-            unscaled_tex = owner.AllocateSurfaceTexture(tuple, rect.GetWidth(), rect.GetHeight());
-        }
-        target_tex = unscaled_tex.handle;
-    }
-
-    OpenGLState cur_state = OpenGLState::GetCurState();
-
-    GLuint old_tex = cur_state.texture_units[0].texture_2d;
-    cur_state.texture_units[0].texture_2d = target_tex;
-    cur_state.Apply();
-
-    // Ensure no bad interactions with GL_UNPACK_ALIGNMENT
-    ASSERT(stride * GetBytesPerPixel(pixel_format) % 4 == 0);
-    if (is_custom) {
-        if (res_scale == 1) {
-            texture = owner.AllocateSurfaceTexture(GetFormatTuple(PixelFormat::RGBA8),
-                                                   custom_tex_info.width, custom_tex_info.height);
-            cur_state.texture_units[0].texture_2d = texture.handle;
-            cur_state.Apply();
-        }
-        // always going to be using rgba8
-        glPixelStorei(GL_UNPACK_ROW_LENGTH, static_cast<GLint>(custom_tex_info.width));
-
-        glActiveTexture(GL_TEXTURE0);
-        glTexSubImage2D(GL_TEXTURE_2D, 0, x0, y0, custom_tex_info.width, custom_tex_info.height,
-                        GL_RGBA, GL_UNSIGNED_BYTE, custom_tex_info.tex.data());
-    } else {
-        glPixelStorei(GL_UNPACK_ROW_LENGTH, static_cast<GLint>(stride));
-
-        glActiveTexture(GL_TEXTURE0);
-        glTexSubImage2D(GL_TEXTURE_2D, 0, x0, y0, static_cast<GLsizei>(rect.GetWidth()),
-                        static_cast<GLsizei>(rect.GetHeight()), tuple.format, tuple.type,
-                        &gl_buffer[buffer_offset]);
-    }
-
-    glPixelStorei(GL_UNPACK_ROW_LENGTH, 0);
-    if (Settings::values.dump_textures && !is_custom)
-        DumpTexture(target_tex, tex_hash);
-
-    cur_state.texture_units[0].texture_2d = old_tex;
-    cur_state.Apply();
-
-    if (res_scale != 1) {
-        auto scaled_rect = rect;
-        scaled_rect.left *= res_scale;
-        scaled_rect.top *= res_scale;
-        scaled_rect.right *= res_scale;
-        scaled_rect.bottom *= res_scale;
-        auto from_rect =
-            is_custom ? Common::Rectangle<u32>{0, custom_tex_info.height, custom_tex_info.width, 0}
-                      : Common::Rectangle<u32>{0, rect.GetHeight(), rect.GetWidth(), 0};
-        if (!owner.texture_filterer->Filter(unscaled_tex.handle, from_rect, texture.handle,
-                                            scaled_rect, type, read_fb_handle, draw_fb_handle)) {
-            BlitTextures(unscaled_tex.handle, from_rect, texture.handle, scaled_rect, type,
-                         read_fb_handle, draw_fb_handle);
-        }
-    }
-
-    InvalidateAllWatcher();
-}
-
-MICROPROFILE_DEFINE(OpenGL_TextureDL, "OpenGL", "Texture Download", MP_RGB(128, 192, 64));
-void CachedSurface::DownloadGLTexture(const Common::Rectangle<u32>& rect, GLuint read_fb_handle,
-                                      GLuint draw_fb_handle) {
-    if (type == SurfaceType::Fill) {
-        return;
-    }
-
-    MICROPROFILE_SCOPE(OpenGL_TextureDL);
-
-    if (gl_buffer.empty()) {
-        gl_buffer.resize(width * height * GetBytesPerPixel(pixel_format));
-    }
-
-    OpenGLState state = OpenGLState::GetCurState();
-    OpenGLState prev_state = state;
-    SCOPE_EXIT({ prev_state.Apply(); });
-
-    const FormatTuple& tuple = GetFormatTuple(pixel_format);
-
-    // Ensure no bad interactions with GL_PACK_ALIGNMENT
-    ASSERT(stride * GetBytesPerPixel(pixel_format) % 4 == 0);
-    glPixelStorei(GL_PACK_ROW_LENGTH, static_cast<GLint>(stride));
-    std::size_t buffer_offset =
-        (rect.bottom * stride + rect.left) * GetBytesPerPixel(pixel_format);
-
-    // If not 1x scale, blit scaled texture to a new 1x texture and use that to flush
-    if (res_scale != 1) {
-        auto scaled_rect = rect;
-        scaled_rect.left *= res_scale;
-        scaled_rect.top *= res_scale;
-        scaled_rect.right *= res_scale;
-        scaled_rect.bottom *= res_scale;
-
-        Common::Rectangle<u32> unscaled_tex_rect{0, rect.GetHeight(), rect.GetWidth(), 0};
-        OGLTexture unscaled_tex =
-            owner.AllocateSurfaceTexture(tuple, rect.GetWidth(), rect.GetHeight());
-        BlitTextures(texture.handle, scaled_rect, unscaled_tex.handle, unscaled_tex_rect, type,
-                     read_fb_handle, draw_fb_handle);
-
-        state.texture_units[0].texture_2d = unscaled_tex.handle;
-        state.Apply();
-
-        glActiveTexture(GL_TEXTURE0);
-        if (GLES) {
-            owner.texture_downloader_es->GetTexImage(GL_TEXTURE_2D, 0, tuple.format, tuple.type,
-                                                     rect.GetHeight(), rect.GetWidth(),
-                                                     &gl_buffer[buffer_offset]);
-        } else {
-            glGetTexImage(GL_TEXTURE_2D, 0, tuple.format, tuple.type, &gl_buffer[buffer_offset]);
-        }
-    } else {
-        state.ResetTexture(texture.handle);
-        state.draw.read_framebuffer = read_fb_handle;
-        state.Apply();
-
-        if (type == SurfaceType::Color || type == SurfaceType::Texture) {
-            glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D,
-                                   texture.handle, 0);
-            glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT, GL_TEXTURE_2D,
-                                   0, 0);
-        } else if (type == SurfaceType::Depth) {
-            glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, 0, 0);
-            glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_TEXTURE_2D,
-                                   texture.handle, 0);
-            glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_STENCIL_ATTACHMENT, GL_TEXTURE_2D, 0, 0);
-        } else {
-            glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, 0, 0);
-            glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT, GL_TEXTURE_2D,
-                                   texture.handle, 0);
-        }
-        switch (glCheckFramebufferStatus(GL_FRAMEBUFFER)) {
-        case GL_FRAMEBUFFER_INCOMPLETE_ATTACHMENT:
-            LOG_WARNING(Render_OpenGL, "Framebuffer incomplete attachment");
-            break;
-        case GL_FRAMEBUFFER_INCOMPLETE_DIMENSIONS:
-            LOG_WARNING(Render_OpenGL, "Framebuffer incomplete dimensions");
-            break;
-        case GL_FRAMEBUFFER_INCOMPLETE_MISSING_ATTACHMENT:
-            LOG_WARNING(Render_OpenGL, "Framebuffer incomplete missing attachment");
-            break;
-        case GL_FRAMEBUFFER_UNSUPPORTED:
-            LOG_WARNING(Render_OpenGL, "Framebuffer unsupported");
-            break;
-        }
-        glReadPixels(static_cast<GLint>(rect.left), static_cast<GLint>(rect.bottom),
-                     static_cast<GLsizei>(rect.GetWidth()), static_cast<GLsizei>(rect.GetHeight()),
-                     tuple.format, tuple.type, &gl_buffer[buffer_offset]);
-    }
-
-    glPixelStorei(GL_PACK_ROW_LENGTH, 0);
 }
 
 enum MatchFlags {
@@ -845,11 +298,9 @@ RasterizerCacheOpenGL::RasterizerCacheOpenGL() {
     texture_filterer = std::make_unique<TextureFilterer>(Settings::values.texture_filter_name,
                                                          resolution_scale_factor);
     format_reinterpreter = std::make_unique<FormatReinterpreterOpenGL>();
-    if (GLES)
+    if (GLES) {
         texture_downloader_es = std::make_unique<TextureDownloaderES>(false);
-
-    read_framebuffer.Create();
-    draw_framebuffer.Create();
+    }
 }
 
 RasterizerCacheOpenGL::~RasterizerCacheOpenGL() {
@@ -866,14 +317,15 @@ bool RasterizerCacheOpenGL::BlitSurfaces(const Surface& src_surface,
                                          const Common::Rectangle<u32>& dst_rect) {
     MICROPROFILE_SCOPE(OpenGL_BlitSurface);
 
-    if (!CheckFormatsBlittable(src_surface->pixel_format, dst_surface->pixel_format))
-        return false;
+    if (CheckFormatsBlittable(src_surface->pixel_format, dst_surface->pixel_format)) {
+        dst_surface->InvalidateAllWatcher();
 
-    dst_surface->InvalidateAllWatcher();
+        const Aspect aspect = ToAspect(src_surface->type);
+        return runtime.BlitTextures(src_surface->texture, {aspect, src_rect},
+                                    dst_surface->texture, {aspect, dst_rect});
+    }
 
-    return BlitTextures(src_surface->texture.handle, src_rect, dst_surface->texture.handle,
-                        dst_rect, src_surface->type, read_framebuffer.handle,
-                        draw_framebuffer.handle);
+    return false;
 }
 
 Surface RasterizerCacheOpenGL::GetSurface(const SurfaceParams& params, ScaleMatch match_res_scale,
@@ -1051,47 +503,18 @@ Surface RasterizerCacheOpenGL::GetTextureSurface(const Pica::Texture::TextureInf
             LOG_CRITICAL(Render_OpenGL, "Unsupported mipmap level {}", max_level);
             return nullptr;
         }
-        OpenGLState prev_state = OpenGLState::GetCurState();
-        OpenGLState state;
-        SCOPE_EXIT({ prev_state.Apply(); });
-        auto format_tuple = GetFormatTuple(params.pixel_format);
 
         // Allocate more mipmap level if necessary
         if (surface->max_level < max_level) {
-            state.texture_units[0].texture_2d = surface->texture.handle;
-            state.Apply();
-            glActiveTexture(GL_TEXTURE0);
-            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAX_LEVEL, max_level);
-
-            // If we are using ARB_texture_storage then we've already allocated all of the mipmap
-            // levels
-            if (!GL_ARB_texture_storage) {
-                u32 width, height;
-                if (surface->is_custom) {
-                    width = surface->custom_tex_info.width;
-                    height = surface->custom_tex_info.height;
-                } else {
-                    width = surface->GetScaledWidth();
-                    height = surface->GetScaledHeight();
-                }
-
-                for (u32 level = surface->max_level + 1; level <= max_level; ++level) {
-                    glTexImage2D(GL_TEXTURE_2D, level, format_tuple.internal_format, width >> level,
-                                 height >> level, 0, format_tuple.format, format_tuple.type,
-                                 nullptr);
-                }
-            }
             if (surface->is_custom || !texture_filterer->IsNull()) {
                 // TODO: proper mipmap support for custom textures
-                glGenerateMipmap(GL_TEXTURE_2D);
+                runtime.GenerateMipmaps(surface->texture, max_level);
             }
+
             surface->max_level = max_level;
         }
 
         // Blit mipmaps that have been invalidated
-        state.draw.read_framebuffer = read_framebuffer.handle;
-        state.draw.draw_framebuffer = draw_framebuffer.handle;
-        state.ResetTexture(surface->texture.handle);
         SurfaceParams surface_params = *surface;
         for (u32 level = 1; level <= max_level; ++level) {
             // In PICA all mipmap levels are stored next to each other
@@ -1101,6 +524,7 @@ Surface RasterizerCacheOpenGL::GetTextureSurface(const Pica::Texture::TextureInf
             surface_params.height /= 2;
             surface_params.stride = 0; // reset stride and let UpdateParams re-initialize it
             surface_params.UpdateParams();
+
             auto& watcher = surface->level_watchers[level - 1];
             if (!watcher || !watcher->Get()) {
                 auto level_surface = GetSurface(surface_params, ScaleMatch::Ignore, true);
@@ -1116,25 +540,16 @@ Surface RasterizerCacheOpenGL::GetTextureSurface(const Pica::Texture::TextureInf
                 if (!level_surface->invalid_regions.empty()) {
                     ValidateSurface(level_surface, level_surface->addr, level_surface->size);
                 }
-                state.ResetTexture(level_surface->texture.handle);
-                state.Apply();
+
                 if (!surface->is_custom && texture_filterer->IsNull()) {
-                    glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D,
-                                           level_surface->texture.handle, 0);
-                    glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT,
-                                           GL_TEXTURE_2D, 0, 0);
+                    const auto src_rect = level_surface->GetScaledRect();
+                    const auto dst_rect = surface_params.GetScaledRect();
+                    const Aspect aspect = ToAspect(surface->type);
 
-                    glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D,
-                                           surface->texture.handle, level);
-                    glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT,
-                                           GL_TEXTURE_2D, 0, 0);
-
-                    auto src_rect = level_surface->GetScaledRect();
-                    auto dst_rect = surface_params.GetScaledRect();
-                    glBlitFramebuffer(src_rect.left, src_rect.bottom, src_rect.right, src_rect.top,
-                                      dst_rect.left, dst_rect.bottom, dst_rect.right, dst_rect.top,
-                                      GL_COLOR_BUFFER_BIT, GL_LINEAR);
+                    runtime.BlitTextures(level_surface->texture, {aspect, src_rect},
+                                         surface->texture, {aspect, dst_rect, level});
                 }
+
                 watcher->Validate();
             }
         }
@@ -1203,32 +618,19 @@ const CachedTextureCube& RasterizerCacheOpenGL::GetTextureCube(const TextureCube
     OpenGLState prev_state = OpenGLState::GetCurState();
     SCOPE_EXIT({ prev_state.Apply(); });
 
-    OpenGLState state;
-    state.draw.read_framebuffer = read_framebuffer.handle;
-    state.draw.draw_framebuffer = draw_framebuffer.handle;
-    state.ResetTexture(cube.texture.handle);
-
     for (const Face& face : faces) {
         if (face.watcher && !face.watcher->IsValid()) {
             auto surface = face.watcher->Get();
             if (!surface->invalid_regions.empty()) {
                 ValidateSurface(surface, surface->addr, surface->size);
             }
-            state.ResetTexture(surface->texture.handle);
-            state.Apply();
-            glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D,
-                                   surface->texture.handle, 0);
-            glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT, GL_TEXTURE_2D,
-                                   0, 0);
 
-            glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, face.gl_face,
-                                   cube.texture.handle, 0);
-            glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT, GL_TEXTURE_2D,
-                                   0, 0);
+            const auto src_rect = surface->GetScaledRect();
+            const auto dst_rect = Common::Rectangle<u32>{0, scaled_size, scaled_size, 0};
+            const Aspect aspect = ToAspect(surface->type);
+            runtime.BlitTextures(surface->texture, {aspect, src_rect},
+                                 cube.texture, {aspect, dst_rect});
 
-            auto src_rect = surface->GetScaledRect();
-            glBlitFramebuffer(src_rect.left, src_rect.bottom, src_rect.right, src_rect.top, 0, 0,
-                              scaled_size, scaled_size, GL_COLOR_BUFFER_BIT, GL_LINEAR);
             face.watcher->Validate();
         }
     }
@@ -1329,7 +731,7 @@ SurfaceSurfaceRect_Tuple RasterizerCacheOpenGL::GetFramebufferSurfaces(
 }
 
 Surface RasterizerCacheOpenGL::GetFillSurface(const GPU::Regs::MemoryFillConfig& config) {
-    Surface new_surface = std::make_shared<CachedSurface>(*this);
+    Surface new_surface = std::make_shared<CachedSurface>(*this, runtime);
 
     new_surface->addr = config.GetStartAddress();
     new_surface->end = config.GetEndAddress();
@@ -1458,8 +860,7 @@ void RasterizerCacheOpenGL::ValidateSurface(const Surface& surface, PAddr addr, 
         // Load data from 3DS memory
         FlushRegion(params.addr, params.size);
         surface->LoadGLBuffer(params.addr, params.end);
-        surface->UploadGLTexture(surface->GetSubRect(params), read_framebuffer.handle,
-                                 draw_framebuffer.handle);
+        surface->UploadGLTexture(surface->GetSubRect(params));
         notify_validated(params.GetInterval());
     }
 }
@@ -1508,17 +909,20 @@ bool RasterizerCacheOpenGL::IntervalHasInvalidPixelFormat(SurfaceParams& params,
 bool RasterizerCacheOpenGL::ValidateByReinterpretation(const Surface& surface,
                                                        SurfaceParams& params,
                                                        const SurfaceInterval& interval) {
-    auto [cvt_begin, cvt_end] =
-        format_reinterpreter->GetPossibleReinterpretations(surface->pixel_format);
-    for (auto reinterpreter = cvt_begin; reinterpreter != cvt_end; ++reinterpreter) {
-        PixelFormat format = reinterpreter->first.src_format;
-        params.pixel_format = format;
+    const PixelFormat dst_format = surface->pixel_format;
+    const SurfaceType type = GetFormatType(dst_format);
+    const FormatTuple& tuple = GetFormatTuple(dst_format);
+
+    for (auto& reinterpreter :
+         format_reinterpreter->GetPossibleReinterpretations(surface->pixel_format)) {
+
+        params.pixel_format = reinterpreter->GetSourceFormat();
         Surface reinterpret_surface =
             FindMatch<MatchFlags::Copy>(surface_cache, params, ScaleMatch::Ignore, interval);
 
         if (reinterpret_surface != nullptr) {
-            SurfaceInterval reinterpret_interval = params.GetCopyableInterval(reinterpret_surface);
-            SurfaceParams reinterpret_params = surface->FromInterval(reinterpret_interval);
+            auto reinterpret_interval = params.GetCopyableInterval(reinterpret_surface);
+            auto reinterpret_params = surface->FromInterval(reinterpret_interval);
             auto src_rect = reinterpret_surface->GetScaledSubRect(reinterpret_params);
             auto dest_rect = surface->GetScaledSubRect(reinterpret_params);
 
@@ -1527,29 +931,30 @@ bool RasterizerCacheOpenGL::ValidateByReinterpretation(const Surface& surface,
                 // The destination surface is either a framebuffer, or a filtered texture.
                 // Create an intermediate surface to convert to before blitting to the
                 // destination.
-                Common::Rectangle<u32> tmp_rect{0, dest_rect.GetHeight() / resolution_scale_factor,
-                                                dest_rect.GetWidth() / resolution_scale_factor, 0};
-                OGLTexture tmp_tex = AllocateSurfaceTexture(
-                    GetFormatTuple(reinterpreter->first.dst_format), tmp_rect.right, tmp_rect.top);
-                reinterpreter->second->Reinterpret(reinterpret_surface->texture.handle, src_rect,
-                                                   read_framebuffer.handle, tmp_tex.handle,
-                                                   tmp_rect, draw_framebuffer.handle);
+                const u32 width = dest_rect.GetHeight() / resolution_scale_factor;
+                const u32 height = dest_rect.GetWidth() / resolution_scale_factor;
+                const Common::Rectangle<u32> tmp_rect{0, width, height, 0};
 
-                const SurfaceType type = GetFormatType(reinterpreter->first.dst_format);
-                if (!texture_filterer->Filter(tmp_tex.handle, tmp_rect, surface->texture.handle,
-                                              dest_rect, type, read_framebuffer.handle,
-                                              draw_framebuffer.handle)) {
-                    BlitTextures(tmp_tex.handle, tmp_rect, surface->texture.handle, dest_rect, type,
-                                 read_framebuffer.handle, draw_framebuffer.handle);
+                OGLTexture tmp_tex = AllocateSurfaceTexture(tuple, height, width);
+                reinterpreter->Reinterpret(reinterpret_surface->texture, src_rect,
+                                            tmp_tex, tmp_rect);
+
+                if (!texture_filterer->Filter(tmp_tex, tmp_rect,
+                                              surface->texture, dest_rect, type)) {
+
+                    const Aspect aspect = ToAspect(type);
+                    runtime.BlitTextures(tmp_tex, {aspect, tmp_rect},
+                                         surface->texture, {aspect, dest_rect});
                 }
             } else {
-                reinterpreter->second->Reinterpret(reinterpret_surface->texture.handle, src_rect,
-                                                   read_framebuffer.handle, surface->texture.handle,
-                                                   dest_rect, draw_framebuffer.handle);
+                reinterpreter->Reinterpret(reinterpret_surface->texture, src_rect,
+                                           surface->texture, dest_rect);
             }
+
             return true;
         }
     }
+
     return false;
 }
 
@@ -1601,9 +1006,9 @@ void RasterizerCacheOpenGL::FlushRegion(PAddr addr, u32 size, Surface flush_surf
 
         if (surface->type != SurfaceType::Fill) {
             SurfaceParams params = surface->FromInterval(interval);
-            surface->DownloadGLTexture(surface->GetSubRect(params), read_framebuffer.handle,
-                                       draw_framebuffer.handle);
+            surface->DownloadGLTexture(surface->GetSubRect(params));
         }
+
         surface->FlushGLBuffer(boost::icl::first(interval), boost::icl::last_next(interval));
         flushed_intervals += interval;
     }
@@ -1680,7 +1085,7 @@ void RasterizerCacheOpenGL::InvalidateRegion(PAddr addr, u32 size, const Surface
 }
 
 Surface RasterizerCacheOpenGL::CreateSurface(const SurfaceParams& params) {
-    Surface surface = std::make_shared<CachedSurface>(*this);
+    Surface surface = std::make_shared<CachedSurface>(*this, runtime);
     static_cast<SurfaceParams&>(*surface) = params;
 
     surface->invalid_regions.insert(surface->GetInterval());
