+++ src/video_core/rasterizer_cache/rasterizer_cache.h
@@ -4,11 +4,9 @@
 
 #pragma once
 #include <unordered_map>
-#include "common/assert.h"
-#include "core/custom_tex_cache.h"
+#include "video_core/rasterizer_cache/cached_surface.h"
 #include "video_core/rasterizer_cache/rasterizer_cache_utils.h"
 #include "video_core/rasterizer_cache/surface_params.h"
-#include "video_core/renderer_opengl/gl_resource_manager.h"
 #include "video_core/texture/texture_decode.h"
 
 namespace OpenGL {
@@ -17,129 +15,6 @@ enum class ScaleMatch {
     Exact,   // only accept same res scale
     Upscale, // only allow higher scale than params
     Ignore   // accept every scaled res
-};
-
-/**
- * A watcher that notifies whether a cached surface has been changed. This is useful for caching
- * surface collection objects, including texture cube and mipmap.
- */
-struct SurfaceWatcher {
-public:
-    explicit SurfaceWatcher(std::weak_ptr<CachedSurface>&& surface) : surface(std::move(surface)) {}
-
-    /**
-     * Checks whether the surface has been changed.
-     * @return false if the surface content has been changed since last Validate() call or has been
-     * destroyed; otherwise true
-     */
-    bool IsValid() const {
-        return !surface.expired() && valid;
-    }
-
-    /// Marks that the content of the referencing surface has been updated to the watcher user.
-    void Validate() {
-        ASSERT(!surface.expired());
-        valid = true;
-    }
-
-    /// Gets the referencing surface. Returns null if the surface has been destroyed
-    Surface Get() const {
-        return surface.lock();
-    }
-
-private:
-    friend struct CachedSurface;
-    std::weak_ptr<CachedSurface> surface;
-    bool valid = false;
-};
-
-class RasterizerCacheOpenGL;
-
-struct CachedSurface : SurfaceParams, std::enable_shared_from_this<CachedSurface> {
-    CachedSurface(RasterizerCacheOpenGL& owner) : owner{owner} {}
-    ~CachedSurface();
-
-    bool CanFill(const SurfaceParams& dest_surface, SurfaceInterval fill_interval) const;
-    bool CanCopy(const SurfaceParams& dest_surface, SurfaceInterval copy_interval) const;
-
-    bool IsRegionValid(SurfaceInterval interval) const {
-        return (invalid_regions.find(interval) == invalid_regions.end());
-    }
-
-    bool IsSurfaceFullyInvalid() const {
-        auto interval = GetInterval();
-        return *invalid_regions.equal_range(interval).first == interval;
-    }
-
-    bool registered = false;
-    SurfaceRegions invalid_regions;
-
-    u32 fill_size = 0; /// Number of bytes to read from fill_data
-    std::array<u8, 4> fill_data;
-
-    OGLTexture texture;
-
-    /// max mipmap level that has been attached to the texture
-    u32 max_level = 0;
-    /// level_watchers[i] watches the (i+1)-th level mipmap source surface
-    std::array<std::shared_ptr<SurfaceWatcher>, 7> level_watchers;
-
-    bool is_custom = false;
-    Core::CustomTexInfo custom_tex_info;
-
-    std::vector<u8> gl_buffer;
-
-    // Read/Write data in 3DS memory to/from gl_buffer
-    void LoadGLBuffer(PAddr load_start, PAddr load_end);
-    void FlushGLBuffer(PAddr flush_start, PAddr flush_end);
-
-    // Custom texture loading and dumping
-    bool LoadCustomTexture(u64 tex_hash);
-    void DumpTexture(GLuint target_tex, u64 tex_hash);
-
-    // Upload/Download data in gl_buffer in/to this surface's texture
-    void UploadGLTexture(Common::Rectangle<u32> rect, GLuint read_fb_handle, GLuint draw_fb_handle);
-    void DownloadGLTexture(const Common::Rectangle<u32>& rect, GLuint read_fb_handle,
-                           GLuint draw_fb_handle);
-
-    std::shared_ptr<SurfaceWatcher> CreateWatcher() {
-        auto watcher = std::make_shared<SurfaceWatcher>(weak_from_this());
-        watchers.push_front(watcher);
-        return watcher;
-    }
-
-    void InvalidateAllWatcher() {
-        for (const auto& watcher : watchers) {
-            if (auto locked = watcher.lock()) {
-                locked->valid = false;
-            }
-        }
-    }
-
-    void UnlinkAllWatcher() {
-        for (const auto& watcher : watchers) {
-            if (auto locked = watcher.lock()) {
-                locked->valid = false;
-                locked->surface.reset();
-            }
-        }
-        watchers.clear();
-    }
-
-private:
-    RasterizerCacheOpenGL& owner;
-    std::list<std::weak_ptr<SurfaceWatcher>> watchers;
-};
-
-struct CachedTextureCube {
-    OGLTexture texture;
-    u16 res_scale = 1;
-    std::shared_ptr<SurfaceWatcher> px;
-    std::shared_ptr<SurfaceWatcher> nx;
-    std::shared_ptr<SurfaceWatcher> py;
-    std::shared_ptr<SurfaceWatcher> ny;
-    std::shared_ptr<SurfaceWatcher> pz;
-    std::shared_ptr<SurfaceWatcher> nz;
 };
 
 class TextureDownloaderES;
@@ -233,13 +108,11 @@ private:
     /// Increase/decrease the number of surface in pages touching the specified region
     void UpdatePagesCachedCount(PAddr addr, u32 size, int delta);
 
+    TextureRuntime runtime;
     SurfaceCache surface_cache;
     PageMap cached_pages;
     SurfaceMap dirty_regions;
     SurfaceSet remove_surfaces;
-
-    OGLFramebuffer read_framebuffer;
-    OGLFramebuffer draw_framebuffer;
 
     u16 resolution_scale_factor;
