+++ src/video_core/renderer_opengl/texture_filters/texture_filterer.cpp
@@ -58,16 +58,16 @@ bool TextureFilterer::IsNull() const {
     return !filter;
 }
 
-bool TextureFilterer::Filter(GLuint src_tex, const Common::Rectangle<u32>& src_rect, GLuint dst_tex,
-                             const Common::Rectangle<u32>& dst_rect,
-                             SurfaceType type, GLuint read_fb_handle,
-                             GLuint draw_fb_handle) {
-    // depth / stencil texture filtering is not supported for now
+bool TextureFilterer::Filter(const OGLTexture& src_tex, Common::Rectangle<u32> src_rect,
+                             const OGLTexture& dst_tex, Common::Rectangle<u32> dst_rect,
+                             SurfaceType type) {
+
+    // Depth/Stencil texture filtering is not supported for now
     if (IsNull() || (type != SurfaceType::Color && type != SurfaceType::Texture)) {
         return false;
     }
 
-    filter->Filter(src_tex, src_rect, dst_tex, dst_rect, read_fb_handle, draw_fb_handle);
+    filter->Filter(src_tex, src_rect, dst_tex, dst_rect);
     return true;
 }
 
@@ -83,6 +83,7 @@ std::vector<std::string_view> TextureFilterer::GetFilterNames() {
             return lhs_is_none && !rhs_is_none;
         return lhs < rhs;
     });
+
     return ret;
 }
