+++ src/video_core/renderer_opengl/texture_filters/xbrz/xbrz_freescale.h
@@ -3,7 +3,6 @@
 // Refer to the license.txt file included.
 
 #pragma once
-
 #include "video_core/renderer_opengl/gl_resource_manager.h"
 #include "video_core/renderer_opengl/gl_state.h"
 #include "video_core/renderer_opengl/texture_filters/texture_filter_base.h"
@@ -15,9 +14,8 @@ public:
     static constexpr std::string_view NAME = "xBRZ freescale";
 
     explicit XbrzFreescale(u16 scale_factor);
-    void Filter(GLuint src_tex, const Common::Rectangle<u32>& src_rect, GLuint dst_tex,
-                const Common::Rectangle<u32>& dst_rect, GLuint read_fb_handle,
-                GLuint draw_fb_handle) override;
+    void Filter(const OGLTexture& src_tex, Common::Rectangle<u32> src_rect,
+                const OGLTexture& dst_tex, Common::Rectangle<u32> dst_rect) override;
 
 private:
     OpenGLState state{};
